Added some styling to steps. Fixed some formset issues for empty forms.

+++ mysite/recipes/views.py
@@ -3,13 +3,11 @@ from urllib import quote_plus
 from django.contrib.auth.decorators import login_required
 from django.core.paginator import Paginator, EmptyPage, PageNotAnInteger
 from django.forms import formset_factory
-from django.forms import modelformset_factory
 from django.http import HttpResponseRedirect, Http404
 from django.shortcuts import render
 
-from .forms import RecipeForm, StepForm, IngredientForm, BaseIngredientFormSet
+from .forms import RecipeForm, StepForm, IngredientForm, BaseIngredientFormSet, BaseStepsFormSet
 from .models import Recipe
-from .models import Step
 
 
 def index(request):
@@ -80,39 +78,45 @@ def add_recipe(request):
 @login_required(login_url='/login/')
 def edit(request, rid):
     edit_recipe = Recipe.objects.get(pk=rid)
-    StepFormSet = modelformset_factory(Step, form=StepForm)
 
+    StepFormSet = formset_factory(StepForm, formset=BaseStepsFormSet)
     IngredientFormSet = formset_factory(IngredientForm, formset=BaseIngredientFormSet)
 
     ingredients = edit_recipe.ingredient_set.all()
     ingredient_data = [{'name': i.name, 'quantity_type': i.quantity_type, 'quantity': i.quantity} for i in ingredients]
 
-    # ingredients_form = IngredientFormSet(queryset=Ingredient.objects.filter(rid=rid))
-    # steps_form = StepFormSet(queryset=Step.objects.filter(rid=rid))
+    steps = edit_recipe.step_set.all()
+    steps_data = [{'description': s.description, 'order': s.order} for s in steps]
 
     if request.POST:
         recipe_form = RecipeForm(request.POST, instance=edit_recipe)
 
         ingredient_formset = IngredientFormSet(request.POST, prefix='ingr')
-        # steps_form = StepFormSet(request.POST, queryset=Step.objects.filter(rid=rid))
+        steps_formset = StepFormSet(request.POST, prefix='steps')
 
-        if recipe_form.is_valid() and ingredient_formset.is_valid():
+        if recipe_form.is_valid() and ingredient_formset.is_valid() and steps_formset.is_valid():
             recipe = recipe_form.save()
-
             recipe.ingredient_set.all().delete()
+            recipe.step_set.all().delete()
 
             for ingr_form in ingredient_formset.forms:
-                ingr = ingr_form.save(commit=False)
-                ingr.rid_id = recipe.rid
-                ingr.save()
-                # if steps_form.is_valid():
-                # steps_form.save()
+                if ingr_form.is_valid() and ingr_form.has_changed():
+                    ingr = ingr_form.save(commit=False)
+                    ingr.rid_id = recipe.rid
+                    ingr.save()
+
+            for steps_form in steps_formset.forms:
+                if steps_form.is_valid() and steps_form.has_changed():
+                    step = steps_form.save(commit=False)
+                    step.rid_id = recipe.rid
+                    step.save()
+
         return HttpResponseRedirect('/recipes/' + str(rid))
     else:
         recipe_form = RecipeForm(instance=edit_recipe)
         ingredients_formset = IngredientFormSet(initial=ingredient_data, prefix='ingr')
-        steps_form = StepFormSet(queryset=Step.objects.filter(rid=rid))
-        context = {'recipe_form': recipe_form, 'rid': rid, 'ingredients_form': ingredients_formset, 'steps_form': steps_form}
+        steps_formset = StepFormSet(initial=steps_data, prefix='steps')
+        context = {'recipe_form': recipe_form, 'rid': rid, 'ingredients_form': ingredients_formset, 'steps_formset': steps_formset}
 
     return render(request, 'recipes/edit.html', context)
